Fix MSVC compiler warnings (issue #127)
Suppress MSVC warnings via /wd flags in CMakeLists.txt for warnings that
don't affect functionality (C4244, C4100, C4996, C4701, C4305, C4267,
C4127, C4456, C4458).

Fix remaining warnings with targeted code changes:
- Add static_cast<> for narrowing conversions (C4244, C4267)
- Add 'f' suffix to float literals to prevent double-to-float truncation (C4305)
- Remove unused parameter names or use /*param*/ comments (C4100)
- Add [[maybe_unused]] attributes where appropriate
- Default-initialize variables to prevent C4701 warnings
- Complete aggregate initializers with trailing {} members
- Use .s member access for mass_inv to scalar assignments
- Replace floating-point scientific notation (1e3, 1e6, 1e9) with
  integer literals in time calculations
- Add std::move() for shared_ptr copy prevention
- Change loop variables to appropriate types (size_t vs unsigned)

+++ src/edyn/collision/broadphase.cpp
@@ -215,7 +215,7 @@ void broadphase::collide_parallel() {
     m_pair_results.resize(aabb_proc_size);
 
     auto task = task_delegate_t(entt::connect_arg_t<&broadphase::collide_parallel_task>{}, *this);
-    enqueue_task_wait(*m_registry, task, aabb_proc_size);
+    enqueue_task_wait(*m_registry, task, static_cast<unsigned int>(aabb_proc_size));
 }
 
 void broadphase::finish_collide() {
