+++ src/edyn/shapes/convex_mesh.cpp
@@ -149,7 +149,7 @@ void convex_mesh::calculate_edges() {
                     // edge will be directed along the circumference of the first face
                     // as well.
                     EDYN_ASSERT(edge_faces[edge_idx * 2 + 1] == std::numeric_limits<uint32_t>::max());
-                    edge_faces[edge_idx * 2 + 1] = face_idx;
+                    edge_faces[edge_idx * 2 + 1] = static_cast<uint32_t>(face_idx);
                     edge_normals[edge_idx * 2 + 1] = normals[face_idx];
                     break;
                 }
@@ -158,7 +158,7 @@ void convex_mesh::calculate_edges() {
             if (!contains) {
                 edges.push_back(vertex_idx0);
                 edges.push_back(vertex_idx1);
-                edge_faces.push_back(face_idx);
+                edge_faces.push_back(static_cast<uint32_t>(face_idx));
                 edge_faces.push_back(std::numeric_limits<uint32_t>::max());
 
                 edge_vertices.push_back(vertices[vertex_idx0]);
@@ -202,7 +202,7 @@ void convex_mesh::calculate_relevant_faces() {
         });
 
         if (found_it == relevant_faces.end()) {
-            relevant_faces.push_back(face_idx);
+            relevant_faces.push_back(static_cast<uint32_t>(face_idx));
             relevant_normals.push_back(normals[face_idx]);
         }
     }
@@ -221,7 +221,7 @@ void convex_mesh::calculate_relevant_edges() {
         });
 
         if (found_it == relevant_edges.end()) {
-            relevant_edges.push_back(edge_idx);
+            relevant_edges.push_back(static_cast<uint32_t>(edge_idx));
         }
     }
 }
